requirements.txt
streamlit
langchain
langchain_groq
pytesseract
torch
transformers
Pillow
PyMuPDF
huggingface_hub
